Refactor env management and update Docker setup
Removed root .env file and updated Dockerfiles and compose files to use service-specific environment files. Improved documentation to clarify service architecture and frontend startup. Refactored frontend toast imports to use the new path. Added Suspense fallback to user dashboard layout for better loading experience.
