--- .env
@@ -1,14 +1,0 @@
-NEXT_PUBLIC_GOOGLE_MAPS_API_KEY=AIzaSyCKk52drd9VMWnrONtrxQgiLctEKny8hwg
-
-NEXT_PUBLIC_SUPABASE_URL=https://cvmxfwmcaxmqnhmsxicu.supabase.co
-NEXT_PUBLIC_SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImN2bXhmd21jYXhtcW5obXN4aWN1Iiwicm9sZSI6ImFub24iLCJpYXQiOjE3NTM4OTkyNjcsImV4cCI6MjA2OTQ3NTI2N30.upgtFa_ZxotPJcf_X6RQFNMxoyuZTfTniNwSHOwjJCY
-
-NODE_ENV=development
-
-
-NEXT_PUBLIC_API_GATEWAY_URL=http://localhost:3000
-
-NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME=dacknfqtw
-NEXT_PUBLIC_CLOUDINARY_API_KEY=157459596551646
-NEXT_PUBLIC_CLOUDINARY_API_SECRET=6j2YIB3XAWWvRDAE2JTHYfloajs
-NEXT_PUBLIC_CLOUDINARY_UPLOAD_PRESET=fitnest_uploads
