+++ backend/AdminService/Dockerfile
@@ -13,8 +13,8 @@ RUN npm install
 # Copy the rest of the application code
 COPY . .
 
-# Copy environment file
-COPY .env ./
+# Copy environment file (if it exists)
+COPY .env* ./
 
 # Create a non-root user for security
 RUN addgroup -g 1001 -S nodejs
