+++ frontend/app/auth/complete-profile/gym/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 import { useState, useEffect } from "react"
-import {useToast} from "@/hooks/use-toast"
+import {useToast} from "@/components/ui/use-toast"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import { Input } from "@/components/ui/input"
