+++ frontend/app/dashboard/trainer/_components/CreatePlan.tsx
@@ -7,7 +7,7 @@ import { Label } from "@/components/ui/label";
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "@/components/ui/select";
 import { Plus, Upload, X } from 'lucide-react';
 import { AddSession, uploadToCloudinary } from '@/lib/api';
-import { useToast } from "@/hooks/use-toast";
+import { useToast } from "@/components/ui/use-toast";
 import { useAuth } from '@/contexts/AuthContext';
 import VerifiedActions from '@/components/VerifiedActions';
