+++ frontend/app/dashboard/user/layout.tsx
@@ -1,11 +1,10 @@
 "use client"
-import React from 'react'
+import React, { Suspense } from 'react'
 import { useSearchParams } from 'next/navigation'
 import TopBar  from './_components/TopBar'
 import { UserDataProvider } from './context/UserContext'
 
-
-function Userlayout({ children }: { children: React.ReactNode }) {
+function LayoutContent({ children }: { children: React.ReactNode }) {
   const searchParams = useSearchParams()
   const isAdminView = searchParams.get('adminView') === 'true'
   
@@ -18,6 +17,18 @@ function Userlayout({ children }: { children: React.ReactNode }) {
           </div>
       </div>
     </UserDataProvider>
+  )
+}
+
+function Userlayout({ children }: { children: React.ReactNode }) {
+  return (
+    <Suspense fallback={
+      <div className="min-h-screen bg-black flex items-center justify-center">
+        <div className="text-white">Loading...</div>
+      </div>
+    }>
+      <LayoutContent>{children}</LayoutContent>
+    </Suspense>
   )
 }
